requirements.txt
# Core ML and Embedding Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Code Analysis Dependencies
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-ruby>=0.20.0
gitignore-parser>=0.1.0

# MCP Server Dependencies
mcp>=0.6.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# CLI and Configuration
click>=8.1.0
pyyaml>=6.0.0

# System Utilities
psutil>=5.9.0